Build model configuration directories

Configure an ensemble of lake model simulations from basic set of inputs.

Usage

build_ensemble(
  aeme_data = NULL,
  config = NULL,
  model = c("dy_cd", "glm_aed", "gotm_wet"),
  mod_ctrls,
  inf_factor = c(glm_aed = 1, dy_cd = 1, gotm_wet = 1),
  outf_factor = c(glm_aed = 1, dy_cd = 1, gotm_wet = 1),
  ext_elev = 0,
  use_bgc = TRUE,
  calc_wbal = TRUE,
  calc_wlev = TRUE,
  use_lw = TRUE,
  coeffs = NULL,
  hum_type = 3,
  path = "."
)

Arguments

aeme_data

aeme; data object.

config

list; loaded via `config <- yaml::read_yaml("aeme.yaml")`

model

vector; of models to be used. Can be `dy_cd`, `glm_aed`, `gotm_wet`.

mod_ctrls

dataframe; of configuration loaded from "mod_ctrls.csv".

inf_factor

vector; containing numeric factor to multiple the inflows. Needs to be named according to the model.

outf_factor

vector; containing numeric factor to multiple the outflows. Needs to be named according to the model.

ext_elev

numeric; extension in elevation for the hypsogrph in metres.

use_bgc

boolean; switch to use the biogeochemical model.

calc_wbal

boolean; calculate water balance.

calc_wlev

boolean; calculate water level.

use_lw

boolean; use incoming longwave radiation. Only applies to GLM-AED.

coeffs

numeric vector of length two; to be used to estimate surface water temperature for estimating evaporation. Defaults to NULL. If water temperature observations are included in `aeme` object, then it will use those to build a linear relationship between air temperature and water temperature. Otherwise. it uses the simple estimation \(temp_water = 5 + 0.75 * temp_air\) from Stefan & Preud'homme, 2007: www.doi.org/10.1111/j.1752-1688.1993.tb01502.x

hum_type

numeric; GOTM humidity metric [1=relative humidity ( 2=wet-bulb temperature, 3=dew point temperature, 4=specific humidity (kg/kg)] Default = 3.

path

filepath; where input files are located relative to `config`.

Value

builds the model ensemble configuration.

aeme object

Examples

install.packages("configr", repos = "http://cran.us.r-project.org")
#> Installing package into ‘/tmp/RtmpBNEadW/temp_libpath1be56419720c’
#> (as ‘lib’ is unspecified)
#> also installing the dependencies ‘ini’, ‘RcppTOML’
tmpdir <- tempdir()
aeme_dir <- system.file("extdata/lake/", package = "AEME")
# Copy files from package into tempdir
file.copy(aeme_dir, tmpdir, recursive = TRUE)
#> [1] TRUE
path <- file.path(tmpdir, "lake")
aeme_data <- yaml_to_aeme(path = path, "aeme.yaml")
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is FALSE
#> Warning: Lake area [152343 m2] is different to the area calculated from the lake
#> shape [152433.09 m2].
mod_ctrls <- read.csv(file.path(path, "model_controls.csv"))
inf_factor = c("glm_aed" = 1)
outf_factor = c("glm_aed" = 1)
model <- c("glm_aed")
build_ensemble(path = path, aeme_data = aeme_data, model = model,
               mod_ctrls = mod_ctrls, inf_factor = inf_factor, ext_elev = 5,
               use_bgc = FALSE, use_lw = TRUE)
#> Building simulation for Wainamu [2023-11-07 22:54:27.303815]
#> Using observed water level
#> Error in water_balance(aeme_time = aeme_time, hyps = hyps, inf = inf,     outf = outf[["outflow"]], obs_lvl = aeme_obs[["level"]],     obs_lake = aeme_obs[["lake"]], obs_met = met, ext_elev = ext_elev,     elevation = elevation, print_plots = FALSE, coeffs = coeffs): object 'model' not found