Estimate lake water balance based on a minimal set of inputs

Estimate lake water balance based on a minimal set of inputs

Usage

calc_wbal(
  aeme_time,
  model,
  use,
  hyps,
  inf,
  outf = NULL,
  level = NULL,
  obs_lake = NULL,
  obs_met,
  ext_elev,
  elevation,
  print_plots = FALSE,
  coeffs = NULL
)

Arguments

aeme_time

list; time object from aeme_object using `time()`

model

vector; of models to be used. Can be `dy_cd`, `glm_aed`, `gotm_wet`.

hyps

data frame of hypsographic curve, elevation (masl) and planar area (m^2)

inf

list of inflow data frames

outf

list of outflow data frames. Default = NULL

level

data frame of lake water level observations.. cols = Date, value

obs_lake

data frame of lake observations in ensemble standard format

obs_met

data frame of meteorology, must include MET_tmpair, MET_wndspd & MET_prvapr, continuous Date, extent defines output extent

ext_elev

numeric; elevation to extend bathymetry

elevation

numeric; elevation of lake

print_plots

logical; print plots of water balance components

coeffs

numeric vector; coefficients for estimating lake surface temperature. Default = NULL

method

character; method for estimating lake water level. Options are "use_obs", "sinusoidal", "constant".

Value

data frame of water balance components which are: - Date - lvlwtr - HYD_flow -...