Check if a variable is present in the model output

Check if a variable is present in the model output

Usage

check_var_in_output(aeme, model, var_sim, ens_n = 1)

Arguments

aeme

aeme; object.

model

vector; of models to be used. Can be `dy_cd`, `glm_aed`, `gotm_wet`.

Value

A logical matrix indicating if the variable is present in the model