Read an AEME object from files

Read an AEME object from files

Usage

read_aeme_from_files(path)

Arguments

path

Path to the directory containing the AEME files.

Value

An AEME object populated with data from the files.

Examples

aeme_file <- system.file("extdata/aeme.rds", package = "AEME")
aeme <- readRDS(aeme_file)
path <- "test_write"
model_controls <- get_model_controls()
aeme <- build_aeme(path = path, aeme = aeme, model = "glm_aed",
model_controls = model_controls)
#> Created missing directory: D:\a\AEME\AEME\docs\reference\test_write
#> ℹ Using observed water level
#> ! Missing values in observed water level
#> ℹ Insufficient water level observations. Using constant water level
#> ℹ Correcting water balance using estimated outflows (method = 2).
#> ℹ Calculating lake level using lake depth and a sinisoidal function.
#> ℹ Building GLM-AED2 for lake wainamu
#> ℹ Copied in GLM nml file
#> ✔ GLM nml validation completed - no issues detected.
aeme <- run_aeme(aeme = aeme, model = "glm_aed", path = path)
#> ℹ Running models... (Have you tried parallelizing?) [2025-12-17 23:07:00]
#> → GLM-AED running... [2025-12-17 23:07:00]
#> ! GLM-AED2 run FAILED! [2025-12-17 23:07:01] glm built using gcc version 12.2.0
#> 
#> nDays= 200; timestep= 3600.000000 (s) Maximum lake depth is 13.070000 Depth
#>   where flow will occur over the crest is 13.070000
#> 
#> Wall clock start time : Wed Dec 17 23:07:00 2025
#> 
#> Simulation begins... Running day 2459061, 0.30% of days complete Running day
#>   2459062, 0.60% of days complete Running day 2459063, 0.89% of days complete
#>   Running day 2459064, 1.19% of days complete Running day 2459065, 1.49% of
#>   days complete Running day 2459066, 1.79% of days complete Running day
#>   2459067, 2.08% of days complete Running day 2459068, 2.38% of days complete
#>   Running day 2459069, 2.68% of days complete Running day 2459070, 2.98% of
#>   days complete Running day 2459071, 3.27% of days complete Running day
#>   2459072, 3.57% of days complete Running day 2459073, 3.87% of days complete
#>   Running day 2459074, 4.17% of days complete Running day 2459075, 4.46% of
#>   days complete Running day 2459076, 4.76% of days complete Running day
#>   2459077, 5.06% of days complete Running day 2459078, 5.36% of days complete
#>   Running day 2459079, 5.65% of days complete Running day 2459080, 5.95% of
#>   days complete Running day 2459081, 6.25% of days complete Running day
#>   2459082, 6.55% of days complete Running day 2459083, 6.85% of days complete
#>   Running day 2459084, 7.14% of days complete Running day 2459085, 7.44% of
#>   days complete Running day 2459086, 7.74% of days complete Running day
#>   2459087, 8.04% of days complete Running day 2459088, 8.33% of days complete
#>   Running day 2459089, 8.63% of days complete Running day 2459090, 8.93% of
#>   days complete Running day 2459091, 9.23% of days complete Running day
#>   2459092, 9.52% of days complete Running day 2459093, 9.82% of days complete
#>   Running day 2459094, 10.12% of days complete Running day 2459095, 10.42% of
#>   days complete Running day 2459096, 10.71% of days complete Running day
#>   2459097, 11.01% of days complete Running day 2459098, 11.31% of days complete
#>   Running day 2459099, 11.61% of days complete Running day 2459100, 11.90% of
#>   days complete Running day 2459101, 12.20% of days complete Running day
#>   2459102, 12.50% of days complete Running day 2459103, 12.80% of days complete
#>   Running day 2459104, 13.10% of days complete Running day 2459105, 13.39% of
#>   days complete Running day 2459106, 13.69% of days complete Running day
#>   2459107, 13.99% of days complete Running day 2459108, 14.29% of days complete
#>   Running day 2459109, 14.58% of days complete Running day 2459110, 14.88% of
#>   days complete Running day 2459111, 15.18% of days complete Running day
#>   2459112, 15.48% of days complete Running day 2459113, 15.77% of days complete
#>   Running day 2459114, 16.07% of days complete Running day 2459115, 16.37% of
#>   days complete Running day 2459116, 16.67% of days complete Running day
#>   2459117, 16.96% of days complete Running day 2459118, 17.26% of days complete
#>   Running day 2459119, 17.56% of days complete Running day 2459120, 17.86% of
#>   days complete Running day 2459121, 18.15% of days complete Running day
#>   2459122, 18.45% of days complete Running day 2459123, 18.75% of days complete
#>   Running day 2459124, 19.05% of days complete Running day 2459125, 19.35% of
#>   days complete Running day 2459126, 19.64% of days complete Running day
#>   2459127, 19.94% of days complete Running day 2459128, 20.24% of days complete
#>   Running day 2459129, 20.54% of days complete Running day 2459130, 20.83% of
#>   days complete Running day 2459131, 21.13% of days complete Running day
#>   2459132, 21.43% of days complete Running day 2459133, 21.73% of days complete
#>   Running day 2459134, 22.02% of days complete Running day 2459135, 22.32% of
#>   days complete Running day 2459136, 22.62% of days complete Running day
#>   2459137, 22.92% of days complete Running day 2459138, 23.21% of days complete
#>   Running day 2459139, 23.51% of days complete Running day 2459140, 23.81% of
#>   days complete Running day 2459141, 24.11% of days complete Running day
#>   2459142, 24.40% of days complete Running day 2459143, 24.70% of days complete
#>   Running day 2459144, 25.00% of days complete Running day 2459145, 25.30% of
#>   days complete Running day 2459146, 25.60% of days complete Running day
#>   2459147, 25.89% of days complete Running day 2459148, 26.19% of days complete
#>   Running day 2459149, 26.49% of days complete Running day 2459150, 26.79% of
#>   days complete Running day 2459151, 27.08% of days complete Running day
#>   2459152, 27.38% of days complete Running day 2459153, 27.68% of days complete
#>   Running day 2459154, 27.98% of days complete Running day 2459155, 28.27% of
#>   days complete Running day 2459156, 28.57% of days complete Running day
#>   2459157, 28.87% of days complete Running day 2459158, 29.17% of days complete
#>   Running day 2459159, 29.46% of days complete Running day 2459160, 29.76% of
#>   days complete Running day 2459161, 30.06% of days complete Running day
#>   2459162, 30.36% of days complete Running day 2459163, 30.65% of days complete
#>   Running day 2459164, 30.95% of days complete Running day 2459165, 31.25% of
#>   days complete Running day 2459166, 31.55% of days complete Running day
#>   2459167, 31.85% of days complete Running day 2459168, 32.14% of days complete
#>   Running day 2459169, 32.44% of days complete Running day 2459170, 32.74% of
#>   days complete Running day 2459171, 33.04% of days complete Running day
#>   2459172, 33.33% of days complete Running day 2459173, 33.63% of days complete
#>   Running day 2459174, 33.93% of days complete Running day 2459175, 34.23% of
#>   days complete Running day 2459176, 34.52% of days complete Running day
#>   2459177, 34.82% of days complete Running day 2459178, 35.12% of days complete
#>   Running day 2459179, 35.42% of days complete Running day 2459180, 35.71% of
#>   days complete Running day 2459181, 36.01% of days complete Running day
#>   2459182, 36.31% of days complete Running day 2459183, 36.61% of days complete
#>   Running day 2459184, 36.90% of days complete Running day 2459185, 37.20% of
#>   days complete Running day 2459186, 37.50% of days complete Running day
#>   2459187, 37.80% of days complete Running day 2459188, 38.10% of days complete
#>   Running day 2459189, 38.39% of days complete Running day 2459190, 38.69% of
#>   days complete Running day 2459191, 38.99% of days complete Running day
#>   2459192, 39.29% of days complete Running day 2459193, 39.58% of days complete
#>   Running day 2459194, 39.88% of days complete Running day 2459195, 40.18% of
#>   days complete Running day 2459196, 40.48% of days complete Running day
#>   2459197, 40.77% of days complete Running day 2459198, 41.07% of days complete
#>   Running day 2459199, 41.37% of days complete Running day 2459200, 41.67% of
#>   days complete Running day 2459201, 41.96% of days complete Running day
#>   2459202, 42.26% of days complete Running day 2459203, 42.56% of days complete
#>   Running day 2459204, 42.86% of days complete Running day 2459205, 43.15% of
#>   days complete Running day 2459206, 43.45% of days complete Running day
#>   2459207, 43.75% of days complete Running day 2459208, 44.05% of days complete
#>   Running day 2459209, 44.35% of days complete Running day 2459210, 44.64% of
#>   days complete Running day 2459211, 44.94% of days complete Running day
#>   2459212, 45.24% of days complete Running day 2459213, 45.54% of days complete
#>   Running day 2459214, 45.83% of days complete Running day 2459215, 46.13% of
#>   days complete Running day 2459216, 46.43% of days complete Running day
#>   2459217, 46.73% of days complete Running day 2459218, 47.02% of days complete
#>   Running day 2459219, 47.32% of days complete Running day 2459220, 47.62% of
#>   days complete Running day 2459221, 47.92% of days complete Running day
#>   2459222, 48.21% of days complete Running day 2459223, 48.51% of days complete
#>   Running day 2459224, 48.81% of days complete Running day 2459225, 49.11% of
#>   days complete Running day 2459226, 49.40% of days complete Running day
#>   2459227, 49.70% of days complete Running day 2459228, 50.00% of days complete
#>   Running day 2459229, 50.30% of days complete Running day 2459230, 50.60% of
#>   days complete Running day 2459231, 50.89% of days complete Running day
#>   2459232, 51.19% of days complete Running day 2459233, 51.49% of days complete
#>   Running day 2459234, 51.79% of days complete Running day 2459235, 52.08% of
#>   days complete Running day 2459236, 52.38% of days complete Running day
#>   2459237, 52.68% of days complete Running day 2459238, 52.98% of days complete
#>   Running day 2459239, 53.27% of days complete Running day 2459240, 53.57% of
#>   days complete Running day 2459241, 53.87% of days complete Running day
#>   2459242, 54.17% of days complete Running day 2459243, 54.46% of days complete
#>   Running day 2459244, 54.76% of days complete Running day 2459245, 55.06% of
#>   days complete Running day 2459246, 55.36% of days complete Running day
#>   2459247, 55.65% of days complete Running day 2459248, 55.95% of days complete
#>   Running day 2459249, 56.25% of days complete Running day 2459250, 56.55% of
#>   days complete Running day 2459251, 56.85% of days complete Running day
#>   2459252, 57.14% of days complete Running day 2459253, 57.44% of days complete
#>   Running day 2459254, 57.74% of days complete Running day 2459255, 58.04% of
#>   days complete Running day 2459256, 58.33% of days complete Running day
#>   2459257, 58.63% of days complete Running day 2459258, 58.93% of days complete
#>   Running day 2459259, 59.23% of days complete Running day 2459260, 59.52% of
#>   days complete Running day 2459261, 59.82% of days complete Running day
#>   2459262, 60.12% of days complete Running day 2459263, 60.42% of days complete
#>   Running day 2459264, 60.71% of days complete Running day 2459265, 61.01% of
#>   days complete Running day 2459266, 61.31% of days complete Running day
#>   2459267, 61.61% of days complete Running day 2459268, 61.90% of days complete
#>   Running day 2459269, 62.20% of days complete Running day 2459270, 62.50% of
#>   days complete Running day 2459271, 62.80% of days complete Running day
#>   2459272, 63.10% of days complete Running day 2459273, 63.39% of days complete
#>   Running day 2459274, 63.69% of days complete Running day 2459275, 63.99% of
#>   days complete Running day 2459276, 64.29% of days complete Running day
#>   2459277, 64.58% of days complete Running day 2459278, 64.88% of days complete
#>   Running day 2459279, 65.18% of days complete Running day 2459280, 65.48% of
#>   days complete Running day 2459281, 65.77% of days complete Running day
#>   2459282, 66.07% of days complete Running day 2459283, 66.37% of days complete
#>   Running day 2459284, 66.67% of days complete Running day 2459285, 66.96% of
#>   days complete Running day 2459286, 67.26% of days complete Running day
#>   2459287, 67.56% of days complete Running day 2459288, 67.86% of days complete
#>   Running day 2459289, 68.15% of days complete Running day 2459290, 68.45% of
#>   days complete Running day 2459291, 68.75% of days complete Running day
#>   2459292, 69.05% of days complete Running day 2459293, 69.35% of days complete
#>   Running day 2459294, 69.64% of days complete Running day 2459295, 69.94% of
#>   days complete Running day 2459296, 70.24% of days complete Running day
#>   2459297, 70.54% of days complete Running day 2459298, 70.83% of days complete
#>   Running day 2459299, 71.13% of days complete Running day 2459300, 71.43% of
#>   days complete Running day 2459301, 71.73% of days complete Running day
#>   2459302, 72.02% of days complete Running day 2459303, 72.32% of days complete
#>   Running day 2459304, 72.62% of days complete Running day 2459305, 72.92% of
#>   days complete Running day 2459306, 73.21% of days complete Running day
#>   2459307, 73.51% of days complete Running day 2459308, 73.81% of days complete
#>   Running day 2459309, 74.11% of days complete Running day 2459310, 74.40% of
#>   days complete Running day 2459311, 74.70% of days complete Running day
#>   2459312, 75.00% of days complete Running day 2459313, 75.30% of days complete
#>   Running day 2459314, 75.60% of days complete Running day 2459315, 75.89% of
#>   days complete Running day 2459316, 76.19% of days complete Running day
#>   2459317, 76.49% of days complete Running day 2459318, 76.79% of days complete
#>   Running day 2459319, 77.08% of days complete Running day 2459320, 77.38% of
#>   days complete Running day 2459321, 77.68% of days complete Running day
#>   2459322, 77.98% of days complete Running day 2459323, 78.27% of days complete
#>   Running day 2459324, 78.57% of days complete Running day 2459325, 78.87% of
#>   days complete Running day 2459326, 79.17% of days complete Running day
#>   2459327, 79.46% of days complete Running day 2459328, 79.76% of days complete
#>   Running day 2459329, 80.06% of days complete Running day 2459330, 80.36% of
#>   days complete Running day 2459331, 80.65% of days complete Running day
#>   2459332, 80.95% of days complete Running day 2459333, 81.25% of days complete
#>   Running day 2459334, 81.55% of days complete Running day 2459335, 81.85% of
#>   days complete Running day 2459336, 82.14% of days complete Running day
#>   2459337, 82.44% of days complete Running day 2459338, 82.74% of days complete
#>   Running day 2459339, 83.04% of days complete Running day 2459340, 83.33% of
#>   days complete Running day 2459341, 83.63% of days complete Running day
#>   2459342, 83.93% of days complete Running day 2459343, 84.23% of days complete
#>   Running day 2459344, 84.52% of days complete Running day 2459345, 84.82% of
#>   days complete Running day 2459346, 85.12% of days complete Running day
#>   2459347, 85.42% of days complete Running day 2459348, 85.71% of days complete
#>   Running day 2459349, 86.01% of days complete Running day 2459350, 86.31% of
#>   days complete Running day 2459351, 86.61% of days complete Running day
#>   2459352, 86.90% of days complete Running day 2459353, 87.20% of days complete
#>   Running day 2459354, 87.50% of days complete Running day 2459355, 87.80% of
#>   days complete Running day 2459356, 88.10% of days complete Running day
#>   2459357, 88.39% of days complete Running day 2459358, 88.69% of days complete
#>   Running day 2459359, 88.99% of days complete Running day 2459360, 89.29% of
#>   days complete Running day 2459361, 89.58% of days complete Running day
#>   2459362, 89.88% of days complete Running day 2459363, 90.18% of days complete
#>   Running day 2459364, 90.48% of days complete Running day 2459365, 90.77% of
#>   days complete Running day 2459366, 91.07% of days complete Running day
#>   2459367, 91.37% of days complete Running day 2459368, 91.67% of days complete
#>   Running day 2459369, 91.96% of days complete Running day 2459370, 92.26% of
#>   days complete Running day 2459371, 92.56% of days complete Running day
#>   2459372, 92.86% of days complete Running day 2459373, 93.15% of days complete
#>   Running day 2459374, 93.45% of days complete Running day 2459375, 93.75% of
#>   days complete Running day 2459376, 94.05% of days complete Running day
#>   2459377, 94.35% of days complete Running day 2459378, 94.64% of days complete
#>   Running day 2459379, 94.94% of days complete Running day 2459380, 95.24% of
#>   days complete Running day 2459381, 95.54% of days complete Running day
#>   2459382, 95.83% of days complete Running day 2459383, 96.13% of days complete
#>   Running day 2459384, 96.43% of days complete Running day 2459385, 96.73% of
#>   days complete Running day 2459386, 97.02% of days complete Running day
#>   2459387, 97.32% of days complete Running day 2459388, 97.62% of days complete
#>   Running day 2459389, 97.92% of days complete Running day 2459390, 98.21% of
#>   days complete Running day 2459391, 98.51% of days complete Running day
#>   2459392, 98.81% of days complete Running day 2459393, 99.11% of days complete
#>   Running day 2459394, 99.40% of days complete Running day 2459395, 99.70% of
#>   days complete
#> ✔ Model run complete! [2025-12-17 23:07:01]
write_aeme_to_files(aeme, path)
aeme_path <- get_lake_dir(aeme = aeme, path = path)
aeme2 <- read_aeme_from_files(aeme_path)