Run aquatic model ensemble

Run aquatic model ensemble

Usage

run_aeme(
  aeme_data,
  model,
  return = TRUE,
  mod_ctrls = NULL,
  nlev = NULL,
  verbose = FALSE,
  debug = FALSE,
  timeout = 0,
  parallel = FALSE,
  check_output = FALSE,
  path = "."
)

Arguments

aeme_data

aeme; data object.

model

vector; of models to be used. Can be `dy_cd`, `glm_aed`, `gotm_wet`.

return

logical; return model output within an `aeme` object? Defaults to TRUE.

mod_ctrls

dataframe; of configuration loaded from "mod_ctrls.csv".

nlev

numeric; number of levels to return in model output. If NULL, calculates number of levels based on lake depth as provided within the hypsograph. If the lake is less thatn 3m deep, it outputs every 0.1m, else if it is greater than 3m, it outputs every 0.33m. Defaults to NULL.

verbose

logical; print model output to console. Defaults to FALSE.

debug

logical; write debug log (Only DYRESM). Defaults to FALSE.

timeout

timeout in seconds, ignored if 0. This is a limit for the elapsed time running command in a separate process. Fractions of seconds are ignored.

parallel

logical; run models in parallel. Defaults to FALSE.

check_output

logical; check model output after running? Defaults to FALSE.

path

filepath; where input files are located relative to `config`.

Value

Runs the model

Examples

if (FALSE) {
tmpdir <- tempdir()
aeme_dir <- system.file("extdata/lake/", package = "AEME")
# Copy files from package into tempdir
file.copy(aeme_dir, tmpdir, recursive = TRUE)
path <- file.path(tmpdir, "lake")
aeme_data <- yaml_to_aeme(path = path, "aeme.yaml")
mod_ctrls <- read.csv(file.path(path, "model_controls.csv"))
inf_factor = c("glm_aed" = 1)
outf_factor = c("glm_aed" = 1)
model <- c("glm_aed")
build_ensemble(path = path, aeme_data = aeme_data, model = model,
               mod_ctrls = mod_ctrls, inf_factor = inf_factor, ext_elev = 5,
               use_bgc = TRUE, use_lw = TRUE)
run_aeme(aeme_data = aeme_data, model = model, verbose = TRUE, path = path,
          return = FALSE)
}